+++ src-ui/app/migrations/telemetry-settings.migrations.ts
@@ -1,10 +1,12 @@
 import { cloneDeep, mergeWith } from 'lodash';
 import { TELEMETRY_SETTINGS_DEFAULT, TelemetrySettings } from '../models/telemetry-settings';
-import { info } from 'tauri-plugin-log-api';
+import { error, info } from 'tauri-plugin-log-api';
 import { v4 as uuidv4 } from 'uuid';
+import { message } from '@tauri-apps/api/dialog';
+import { BaseDirectory, writeTextFile } from '@tauri-apps/api/fs';
 
 const migrations: { [v: number]: (data: any) => any } = {
-  1: toLatest,
+  1: resetToLatest,
   2: from1to2,
 };
 
@@ -12,7 +14,7 @@ export function migrateTelemetrySettings(data: any): TelemetrySettings {
   let currentVersion = data.version || 0;
   // Reset to latest when the current version is higher than the latest
   if (currentVersion > TELEMETRY_SETTINGS_DEFAULT.version) {
-    data = toLatest(data);
+    data = resetToLatest(data);
     info(
       `[telemetry-settings-migrations] Reset future telemetry settings version back to version ${
         currentVersion + ''
@@ -20,7 +22,24 @@ export function migrateTelemetrySettings(data: any): TelemetrySettings {
     );
   }
   while (currentVersion < TELEMETRY_SETTINGS_DEFAULT.version) {
-    data = migrations[++currentVersion](data);
+    try {
+      data = migrations[++currentVersion](data);
+    } catch (e) {
+      error(
+        "[telemetry-settings-migrations] Couldn't migrate to version " +
+          currentVersion +
+          '. Backing up configuration and resetting to the latest version. : ' +
+          e
+      );
+      saveBackup(cloneDeep(data));
+      data = resetToLatest(data);
+      currentVersion = data.version;
+      message(
+        'Your telemetry settings could not to be migrated to the new version of OyasumiVR, and have therefore been reset. Apologies for the inconvenience.\n\nPlease report this issue to the developer so this issue may be fixed in the future. Thank you!',
+        { title: 'Migration Error (Telemetry Settings)' }
+      );
+      continue;
+    }
     currentVersion = data.version;
     info(
       `[telemetry-settings-migrations] Migrated telemetry settings to version ${
@@ -36,6 +55,12 @@ export function migrateTelemetrySettings(data: any): TelemetrySettings {
   return data as TelemetrySettings;
 }
 
+async function saveBackup(oldData: any) {
+  await writeTextFile('telemetry-settings.backup.json', JSON.stringify(oldData, null, 2), {
+    dir: BaseDirectory.AppData,
+  });
+}
+
 function from1to2(data: any): any {
   return {
     enabled: data.enabled,
@@ -43,7 +68,7 @@ function from1to2(data: any): any {
   };
 }
 
-function toLatest(data: any): any {
+function resetToLatest(data: any): any {
   // Reset to latest
   const telemetryId = data.telemetryId || uuidv4();
   data = cloneDeep(TELEMETRY_SETTINGS_DEFAULT);